backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
better-exceptions==0.3.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.23.7